Refactor project store
