+++ components/project/grid.vue
@@ -1,15 +1,15 @@
 <script lang="ts" setup>
-  import useProjectStore from '~~/stores/useProjectStore';
+  import useProject from '~~/stores/useProject';
   import { ProjectWithProjectKey } from '~~/types/models';
 
-  const projects = useProjectStore();
+  const projectStore = useProject();
 
   const handleOpen = (project: ProjectWithProjectKey) => {
     /**
       We can directly assign the project to the target here
       to reduce backend requests.
     */
-    projects.target = project;
+    projectStore.target = project;
 
     navigateTo(`/project/${project.project_keys[0].api_key}/apis`);
   };
@@ -18,7 +18,7 @@
 <template>
   <section class="gap-4 grid grid-cols-1 md:grid-cols-2 mt-4">
     <article
-      v-for="project in projects.list"
+      v-for="project in projectStore.list"
       :key="project.id"
       class="card border border-gray-300 hover:bg-gray-300"
       @click="handleOpen(project)"
