+++ stores/index.ts
@@ -1,6 +1,7 @@
 export { default as useApiStore } from './useApiStore';
-export { default as useProjectStore } from './useProjectStore';
 export { default as useProjectKeyStore } from './useProjectKeyStore';
 export { default as useResourceDataStore } from './useResourceDataStore';
 export { default as useResourceDataTypeStore } from './useResourceDataTypeStore';
 export { default as useResourceModelStore } from './useResourceModelStore';
+
+export { default as useProject } from './useProject';
